refactor(cli): keep server entrypoint focused on lifecycle
